refactor-out msgsign.py from auth.py
